Add SMB Volume Driver
Introduces a volume driver which makes use of SMB shares for
hosting volumes as disk images, having a similar workflow with
the other NFS like drivers.

The driver uses Samba in order to access the SMB shares. It takes
SMB shares information such as share path, credentials and other
mount flags from a config file of which location can be configured
in the cinder config file.

It includes all the features required by the Juno release.

Implements: blueprint smbfs-volume-driver

Change-Id: I3b4bebf73122d5b784228af18f9b1eddfc02b643