Add Windows SMB Volume Driver
Introduces a volume driver which makes use of SMB shares for
hosting volumes as disk images, having a similar workflow with
the other NFS like drivers.

This driver is based on the SMB Volume driver proposed for
Linux, overriding the platform specific methods.

It includes all the features required by the Juno release.

Change-Id: Idd5964690bca618c11fd116f80dd802d6cc2f31b
Implements: blueprint windows-smbfs-volume-driver